+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Use `DefaultReactActivityDelegate` instead of `ReactActivityDelegate`
+- Fix compatibility with React Native 71
+
 ## [0.8.0] - 2023-08-25
 
 ### Added